not sure why but motors working now
Previously when running this test, the firmware would seem to crash
after one firmware.run. It seemed related to the number of PWM objects
that were allocated in memory, leading me to believe there was a
problem with mem allocation, perhaps in the linker script. Not sure why
it is working now...